requirements.txt
pymongo==3.4.0
tornado==4.5.2
requests==2.17.3
PyJWT==1.5.2
cryptography==2.0.3